pytorch_training/requirements.txt
h5py==3.9.0
matplotlib==3.8.3
numpy==2.1.3
PyYAML==6.0.2
scikit_learn==1.3.1
torch==2.1.1+cu118
torch_geometric==2.3.1
tqdm==4.65.0
